Today I learned difference between block and inline compinents
As for me, it's so important thing because it will help me in work in future. It's not such a big information, but I have a lot of things to do today. This is step is small, but so important. Tomorrow I will check more information about CSS and HTML. See ya