Learned some new features
Today I was grinding html and css for 5 hours (with breakes) and learn a lot of different stuff. Also I started making my first steps for my own portfolio with html and css - navigation bar. Now it is looks  so strange, but next topic which I will learn are css box and flexbox, so sooner this page will be more beautiful. See ya later